app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.124.4",
    "groq>=0.37.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
    "wsproto>=1.3.2",
]
